requirements.txt
discord.py==1.7.3
python-dotenv==0.20.0
requests=2.27.1